% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{addEnvToRenviron}
\alias{addEnvToRenviron}
\title{addEnvToRenviron}
\usage{
addEnvToRenviron(chosenEnv)
}
\arguments{
\item{chosenEnv}{string. The selected environment}
}
\value{
Nothing. Overwrite .Renviron
}
\description{
add a line to the .Renviron file of the new project to specify
the active environment used by config
}
