% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_rmarkdown_project}
\alias{create_rmarkdown_project}
\title{create_rmarkdown_project}
\usage{
create_rmarkdown_project(path, ...)
}
\arguments{
\item{path}{the path to the new project. Will also serve as project name}

\item{...}{additional arguments}
}
\value{
Nothing. Create a new RStudio project
}
\description{
Main function to create a new project. Triggered by the new
project wizard. It's callback is at:
inst/rstudio/templates/project/create_project.dcf
}
