% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{demoPath_sys}
\alias{demoPath_sys}
\title{demoPath_sys}
\usage{
demoPath_sys(..., lib.loc = NULL, mustWork = FALSE)
}
\arguments{
\item{...}{additional arguments}

\item{lib.loc}{the path to the R library}

\item{mustWork}{logical}
}
\value{
string
}
\description{
return the path where the template is saved locally
}
