% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{replace_package_name}
\alias{replace_package_name}
\title{replace_package_name}
\usage{
replace_package_name(copied_files, package_name, path)
}
\arguments{
\item{copied_files}{all the files to search for replacements (basename())}

\item{package_name}{the new project name}

\item{path}{the full root path of the copied_files}
}
\value{
Nothing. Overwrite files
}
\description{
replace the generic reference to a template "templetedemo"
with the actual name of the new project
}
