% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{replace_word}
\alias{replace_word}
\title{replace_word}
\usage{
replace_word(file, pattern, replace)
}
\arguments{
\item{file}{the path to the file}

\item{pattern}{the pattern to be replaced}

\item{replace}{the replacement}
}
\value{
Nothing. Overwrite the file
}
\description{
replace a pattern in a file
}
